Splitting GameMetadata tests.
